// src/interfaces/ProductsInterface.tsx
// Generated by https://quicktype.io

export interface ProductsInterface {
    total:     number;
    productos: Producto[];
}

export interface Producto {
    precio:    number;
    _id:       string;
    nombre:    string;
    categoria: Categoria;
    usuario:   Categoria;
    img:       string;
}

export interface Categoria {
    _id:    string;
    nombre: string;
}


// Generated by https://quicktype.io

export interface NewProductBody {
    nombre:    string;
    categoria: string;
    precio:    number;
}


// Generated by https://quicktype.io

export interface CategoryResponse {
    total:      number;
    categorias: Categoria[];
}

export interface Categoria {
    _id:     string;
    nombre:  string;
    usuario: Usuario;
}

export interface Usuario {
    _id:    string;
    nombre: string;
}